+++ rspice/lib/rspice/shared_context.rb
@@ -2,4 +2,5 @@
 
 require "rspice/shared_context/with_an_example_descendant_class"
 require "rspice/shared_context/with_callbacks"
+require "rspice/shared_context/with_class_callbacks"
 require "rspice/shared_context/with_example_class_having_callback"